fix: 특정 공고에 대한 지원자 목록 조회 기능 코드 이동
