+++ src/main/java/com/halo/core_bridge/api/jobposting/controller/JobPostingController.java
@@ -7,6 +7,8 @@ import com.halo.core_bridge.api.jobposting.model.dto.JobPostingDto;
 import com.halo.core_bridge.api.jobposting.service.JobPostingService;
 import com.halo.core_bridge.api.jobposting.service.JobPostingSkillService;
 import com.halo.core_bridge.api.jobposting.service.RecruitProcessService;
+import com.halo.core_bridge.api.resume.model.dto.ResumeDto;
+import com.halo.core_bridge.api.resume.service.ResumeService;
 import com.halo.core_bridge.common.model.BaseResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -203,31 +205,6 @@ public class JobPostingController {
         JobPostingDto.HeaderResponse headerDetail = jobPostingService.getHeaderDetail(id);
         return ResponseEntity.ok(BaseResponse.success(headerDetail));
     }
-
-    //해당채용공고의 지원자 조회
-    @Operation(
-            summary = "채용공고 지원자 조회",
-            description = "특정 ID의 지원자를 전체 조회합니다.",
-            responses = {
-                    @ApiResponse(
-                            responseCode = "200",
-                            description = "조회 성공",
-                            content = @Content(
-                                    mediaType = "application/json",
-                                    examples = @ExampleObject(
-                                            name = "Applicants Response Example",
-                                            value = SwaggerJobPostingContents.APPLICANT_RESPONSE
-                                    )
-                            )
-                    )
-            }
-    )
-    @GetMapping("/{id}/applicants")
-    public ResponseEntity<BaseResponse<List<JobPostingDto.ApplicantResponse>>> getJobPostingApplicants(@PathVariable Long id) {
-        List<JobPostingDto.ApplicantResponse> result = jobPostingService.getApplicants(id);
-        return ResponseEntity.ok(BaseResponse.success(result));
-    }
-
 
 
     //채용공고 수정
